requirements.txt
streamlit
openai
python-dotenv
pandas
plotly
pandas_datareader
